+++ src/app/Application.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <memory>
 
-#include "src/gtkmm3/GtkmmIniter.h"
+#include "src/gtkmm4/GtkmmIniter.h"
 #include "src/log/log.h"
 
 namespace app
@@ -19,7 +19,7 @@ int Application::run(std::shared_ptr<ApplicationContext> ctx)
     return INVALID;
   }
 
-  std::shared_ptr<Gtkmm3i::GtkmmIniter> gtkmmIniter = create_gtkmm_initer();
+  std::shared_ptr<Gtkmm4i::GtkmmIniter> gtkmmIniter = create_gtkmm_initer();
 
   if (gtkmmIniter->run(ctx) != 0) {
     LOGE("The gtkmm controller returned invalid execution status");
@@ -29,9 +29,9 @@ int Application::run(std::shared_ptr<ApplicationContext> ctx)
   return 0;
 }
 
-std::shared_ptr<Gtkmm3i::GtkmmIniter> Application::create_gtkmm_initer()
+std::shared_ptr<Gtkmm4i::GtkmmIniter> Application::create_gtkmm_initer()
 {
-  return std::make_shared<Gtkmm3i::GtkmmIniter>();
+  return std::make_shared<Gtkmm4i::GtkmmIniter>();
 }
 
 }  // namespace app