+++ src/app/tests/component/CTEST_app.cpp
@@ -2,6 +2,7 @@
 #include <gtest/gtest.h>
 
 #include "src/app/ApplicationFactory.h"
+#include "src/gtkmm4/GtkmmIniter.h"
 
 using namespace app;
 using namespace testing;