+++ src/gtkmm4/gtkmm4_includes.h
@@ -1,5 +1,5 @@
-#ifndef YOUR_CPP_APP_TEMPLATE_PROJECT_GTKMM_INCLUDES_H
-#define YOUR_CPP_APP_TEMPLATE_PROJECT_GTKMM_INCLUDES_H
+#ifndef YOUR_CPP_APP_TEMPLATE_PROJECT_GTKMM4_INCLUDES_H
+#define YOUR_CPP_APP_TEMPLATE_PROJECT_GTKMM4_INCLUDES_H
 
 #pragma GCC system_header
 #pragma GCC diagnostic push
@@ -9,11 +9,11 @@
 #include <gio/gio.h>
 #include <giomm/resource.h>
 #include <gtkmm.h>
-#include <gtkmm/hvbox.h>
+#include <gtkmm/box.h>
 #include <gtkmm/image.h>
 #include <gtkmm/label.h>
 
 #pragma warning(pop)
 #pragma GCC diagnostic pop
 
-#endif  // YOUR_CPP_APP_TEMPLATE_PROJECT_GTKMM_INCLUDES_H
+#endif  // YOUR_CPP_APP_TEMPLATE_PROJECT_GTKMM4_INCLUDES_H